minor code improvements and refactors

+++ password_generator.c
@@ -1,10 +1,10 @@
 #include "password_generator.h"
 
-char *format_string(char string[], const int string_len)
+char *format_string(char string[], const size_t string_len)
 {
-	for (int i = 0; i < string_len; i++) {
+	for (size_t i = 0; i < string_len; i++) {
 		if (string[i] == ' ') {
-			memmove(&string[i], &string[i+1], string_len - i);
+			memmove(&string[i], &string[i + 1], string_len - i);
 		}
 	}
